requirements.txt
flask==3.0.3
python-nmap==0.7.1
colorama==0.4.6
flask-sqlalchemy==3.1.1